Altered AbstractCamera to work when Layer == null
If `Layer == null`, then the `TargetToScreenTransform` is simply set to
the identity matrix (which makes sense since without a layer, the
"target" *is* the screen). This change prevents `PhysicalObject`s from
throwing `NullReferenceException`s when attempting to convert between
world, target, and screen coordinates when not attached to a layer.